refactor: use reversed Comparator to minimize queue size
Because States are compared by fitness, we should explore the best trains first. With a toy example of 12 dominos, a PriorityQueue that sorts elements in ascending order results in an average queue size 12.6 with a maximum size of 24. But by sorting elements in descending order, the average and maximum drops to 3.7 and 7, respectively.